# Add lookup tables to DiagnosticsManager to speed up diagnostic clearing