… src/DiagnosticFilterer.spec.ts
… src/DiagnosticManager.spec.ts
… src/Scope.ts
… src/bscPlugin/CallExpressionInfo.ts
… src/bscPlugin/validation/ScopeValidator.ts
… src/files/BrsFile.ts
… src/files/XmlFile.spec.ts
… src/interfaces.ts
… src/types/BscType.ts
… src/util.ts