@markwpearce 2025-02-20
Some of the work here is to make the overall memory footprint smaller:

Here's before & after snapshots of the heap memory when validating & building Jellyfin (with an extra file that adds 4000+ namespaces):
<img width="213" alt="image" src="https://github.com/user-attachments/assets/00ddf30d-5950-4737-a4a0-6277262f2693" />